# specs/001-crud-empleados/contracts/empleados.openapi.yaml
openapi: 3.0.3
info:
  title: API CRUD Empleados
  version: 1.0.0
  description: Contrato de endpoints para gestión de empleados con eliminación lógica.
servers:
  - url: http://localhost:8080
security:
  - basicAuth: []
paths:
  /api/empleados:
    get:
      summary: Listar empleados activos
      operationId: listEmpleados
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
          description: Cuando se envía, activa modo paginado con tamaño fijo de 5.
      responses:
        '200':
          description: Lista de empleados activos (completa o paginada)
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/EmpleadoResponse'
                  - $ref: '#/components/schemas/EmpleadoPageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Crear empleado
      operationId: createEmpleado
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmpleadoCreateRequest'
      responses:
        '201':
          description: Empleado creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmpleadoResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Clave duplicada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/empleados/{clave}:
    parameters:
      - name: clave
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Consultar empleado activo por clave
      operationId: getEmpleadoByClave
      responses:
        '200':
          description: Empleado encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmpleadoResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Actualizar empleado activo por clave
      operationId: updateEmpleado
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmpleadoUpdateRequest'
      responses:
        '200':
          description: Empleado actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmpleadoResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Eliminar lógicamente empleado por clave
      operationId: deleteEmpleado
      responses:
        '204':
          description: Empleado marcado como inactivo
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Empleado ya eliminado lógicamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
    EmpleadoCreateRequest:
      type: object
      required: [clave, nombre, telefono]
      properties:
        clave:
          type: string
          minLength: 1
        nombre:
          type: string
          minLength: 1
          maxLength: 100
        telefono:
          type: string
          minLength: 1
          maxLength: 100
    EmpleadoUpdateRequest:
      type: object
      required: [nombre, telefono]
      properties:
        nombre:
          type: string
          minLength: 1
          maxLength: 100
        telefono:
          type: string
          minLength: 1
          maxLength: 100
    EmpleadoResponse:
      type: object
      required: [clave, nombre, telefono, activo]
      properties:
        clave:
          type: string
        nombre:
          type: string
          maxLength: 100
        telefono:
          type: string
          maxLength: 100
        activo:
          type: boolean
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
    EmpleadoPageResponse:
      type: object
      required: [content, page, size, totalElements, totalPages]
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/EmpleadoResponse'
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          enum: [5]
        totalElements:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
  responses:
    BadRequest:
      description: Error de validación
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Credenciales inválidas o ausentes
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Empleado no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'